refactor(api): relocate simple_server example to root examples/ member
Move `crates/api/examples/simple_server.rs` to
`examples/examples/api_simple_server.rs` per house rule (AGENTS.md:
runnable examples live in the root-level `examples/` workspace member).

The relocated example is rewritten as a minimal, honest `nebula-api`
library startup demo: in-memory repos + `InMemoryIdempotencyStore` +
`ApiConfig::from_env()` + `app::build_app` + `app::serve`. The
PG/sqlx/`feature="postgres"` composition that was in the old example
belongs to the `apps/server` composition root, not a library demo.

Dep cleanup in `crates/api/Cargo.toml`: the optional `sqlx` dep and
`dep:sqlx` in the `postgres` feature were only ever used by the deleted
example (grep of `crates/api/src/` shows zero `sqlx` usage; all
`cfg(feature="postgres")` usage is also zero in the library src).
`postgres = ["nebula-storage/postgres"]` is kept — it is a valid
feature forward consumed by `apps/server`.

`deny.toml` wrappers updated to allow `nebula-examples` to depend on
`nebula-api` and `nebula-storage` (the startup demo wires in-memory
repos, mirroring `nebula-server`).
